Variables show up in the equation now
Updated the findVariables function to search for all numbers, rather than a
specific "variable" format.

Equation is generated on the fly, using innerHTML and strings that
represent the divs rather than creating div's through jquery directly.

Needs some more formatting, and a way to specify the rate of change as
variables are dragged (sig-figs?).